Fixed all tests
